--- source/js/modules/intro.js
@@ -1,7 +1,0 @@
-export default () => {
-  document.addEventListener(`DOMContentLoaded`, () => {
-    const introMessage = document.querySelector(`.js-intro-message`);
-    introMessage.classList.add(`show-intro-message`);
-  });
-
-};
