+++ source/scss/blocks/form.scss
@@ -45,7 +45,6 @@
   outline: none;
   position: relative;
   margin-left: auto;
-  transition: width 1s ease;
 
   &:focus {
     border: 1px solid $c-purple;
