+++ source/scss/blocks/screen.scss
@@ -46,6 +46,7 @@
   color: $c-dark;
   background-color: $c-light;
   z-index: 1;
+  transform: translateY(3rem);
 
   @media (max-width: $tablet) and (orientation: portrait) {
     width: 3.5rem;
