+++ source/scss/blocks/social-block.scss
@@ -17,7 +17,6 @@
   height: 4rem;
   color: $c-dark;
   border-radius: 50%;
-  transition: all .2s ease;
 
   @media (max-width: $tablet) and (orientation: portrait) {
     display: none;
@@ -30,13 +29,12 @@
   &::before {
     @include center;
     content: "";
-    width: 100%;
-    height: 100%;
+    width: 0%;
+    height: 0%;
     border-radius: 50%;
     background-color: $c-light;
     border: 2px solid $c-light;
     z-index: -1;
-    transition: all .3s ease;
   }
 
   svg {
@@ -44,16 +42,6 @@
     display: block;
     width: 2.3rem;
     height: 2.2rem;
-  }
-
-  &:hover,
-  .social-block--active & {
-    color: $c-purple;
-
-    &::before {
-      width: 112.5%;
-      height: 112.5%;
-    }
   }
 
   &:focus {
@@ -177,7 +165,6 @@
   border: 2px solid transparent;
   border-radius: 50%;
   opacity: 0;
-  transition: color .6s ease;
 
   @media (min-width: $stop-scaling) {
     width: 40px;
@@ -217,8 +204,6 @@
   }
 
   &--fb {
-    transition: color 0.3s linear, opacity 0.3s linear;
-    transition-delay: 0s, 0.1s;
 
     svg {
       width: 0.8rem;
@@ -232,8 +217,6 @@
   }
 
   &--insta {
-    transition: color 0.3s linear, opacity 0.3s linear;
-    transition-delay: 0s, 0.2s;
 
     svg {
       width: 1.5rem;
@@ -247,8 +230,6 @@
   }
 
   &--vk {
-    transition: color 0.3s linear, opacity 0.3s linear;
-    transition-delay: 0s, 0.3s;
 
     svg {
       width: 1.7rem;
@@ -260,17 +241,4 @@
       }
     }
   }
-}
-
-.social-block--active {
-
-  &:hover {
-
-    .social-block__link--fb,
-    .social-block__link--insta,
-    .social-block__link--vk {
-      opacity: 1;
-    }
-  }
-
 }
