+++ source/scss/style.scss
@@ -20,3 +20,4 @@
 @import "blocks/screen.scss";
 @import "blocks/slider.scss";
 @import "blocks/social-block.scss";
+@import "blocks/animations.scss";
